feature: fixed packaging issues
