+++ client/api/teams/service.ts
@@ -1,8 +1,8 @@
-import { TeamResponse, SeasonType } from "./types";
+import { TeamResponse } from "./types";
 
 const createTeam = async (
   group_id: string,
-  name: string
+  name: string,
 ): Promise<TeamResponse> => {
   try {
     const response = await fetch(`/api/team/create`, {
@@ -20,33 +20,4 @@ const createTeam = async (
   }
 };
 
-const createSeason = async (
-  name: string,
-  start_date: string,
-  end_date: string,
-  is_active: boolean = true,
-  league_id: number = 1
-): Promise<SeasonType> => {
-  try {
-    const response = await fetch("/api/seasons", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        name,
-        start_date,
-        end_date,
-        is_active,
-        league_id,
-      }),
-    });
-    const result = await response.json();
-    return result.data;
-  } catch (error) {
-    console.error("error fetching data:", error);
-    throw error;
-  }
-};
-
-export { createTeam, createSeason };
+export { createTeam };
