+++ client/pages/Home/Home.tsx
@@ -1,10 +1,11 @@
-import Leagues from "../Leagues/Leagues";
+// import Leagues from "../Leagues/Leagues";
+import Seasons from "../Seasons/Seasons";
 import { useAuth } from "../../components/AuthContext/AuthContext";
 
 const Home = () => {
   const { currentUser } = useAuth();
 
-  return <div>{currentUser ? <Leagues /> : <></>}</div>;
+  return <div>{currentUser ? <Seasons /> : <></>}</div>;
 };
 
 export default Home;
