+++ client/pages/Leagues/types.tsx
@@ -4,4 +4,13 @@ interface LeaguesProps {
   currentUser: User;
 }
 
-export type { LeaguesProps };
+interface LeagueType {
+  id: number;
+  group_id: number;
+  name: string;
+  description: string;
+  created_at: Date;
+  updated_at: Date;
+}
+
+export type { LeaguesProps, LeagueType };
